services/mqtt-to-kafka-bridge/requirements.txt
paho-mqtt==1.6.1
confluent-kafka==2.5.3
orjson==3.10.7
prometheus-client==0.20.0
